Add basic WebSecurity logic

+++ src/main/java/com/example/sportsbetting/database/repository/IBetRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsbetting.database.repository;
 import com.example.sportsbetting.database.model.Bet;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IBetRepository extends CrudRepository<Bet, Long> {
+public interface IBetRepository extends CrudRepository<Bet, Integer> {
 }
