+++ src/main/java/com/example/sportsbetting/database/repository/IOutcomeOddRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsbetting.database.repository;
 import com.example.sportsbetting.database.model.OutcomeOdd;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IOutcomeOddRepository extends CrudRepository<OutcomeOdd, Long> {
+public interface IOutcomeOddRepository extends CrudRepository<OutcomeOdd, Integer> {
 }
