+++ src/main/java/com/example/sportsbetting/database/repository/IOutcomeRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsbetting.database.repository;
 import com.example.sportsbetting.database.model.Outcome;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IOutcomeRepository extends CrudRepository<Outcome, Long> {
+public interface IOutcomeRepository extends CrudRepository<Outcome, Integer> {
 }
