+++ src/main/java/com/example/sportsbetting/database/repository/IResultRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsbetting.database.repository;
 import com.example.sportsbetting.database.model.Result;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IResultRepository extends CrudRepository<Result, Long> {
+public interface IResultRepository extends CrudRepository<Result, Integer> {
 }
