+++ src/main/java/com/example/sportsbetting/database/repository/IWagerRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsbetting.database.repository;
 import com.example.sportsbetting.database.model.Wager;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IWagerRepository extends CrudRepository<Wager, Long> {
+public interface IWagerRepository extends CrudRepository<Wager, Integer> {
 }
