+++ src/main/java/com/example/sportsbetting/database/service/BetService.java
@@ -4,7 +4,6 @@ import com.example.sportsbetting.database.model.Bet;
 import com.example.sportsbetting.database.repository.IBetRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-
 import java.util.List;
 
 @Component
@@ -15,7 +14,7 @@ public class BetService {
     public void add(Bet bet) {
         repository.save(bet);
     }
-    public void delete(long id) {
+    public void delete(int id) {
         repository.deleteById(id);
     }
     public List<Bet> getBets() {
